# [DNM] Demo throwing AttestationError performance