--- a/packages/lodestar/test/perf/chain/validation/attestation.test.ts
+++ b/packages/lodestar/test/perf/chain/validation/attestation.test.ts
@@ -4,6 +4,7 @@ import {validateGossipAttestation} from "../../../../src/chain/validation";
 import {generateTestCachedBeaconStateOnlyValidators} from "@chainsafe/lodestar-beacon-state-transition/test/perf/util";
 import {getAttestationValidData} from "../../../utils/validationData/attestation";
 import {AttestationError, AttestationErrorCode, GossipAction} from "../../../../src/chain/errors";
+import {ZERO_HASH} from "../../../../src/constants";
 
 describe("validate gossip attestation", () => {
   const vc = 64;
@@ -30,43 +31,41 @@ describe("validate gossip attestation", () => {
 
 describe.only("test throw AttestationError vs object", () => {
   itBench({
-    id: "AttestationError",
+    id: "AttestationError x1000",
+    runsFactor: 1000,
     fn: async () => {
       for (let i = 0; i < 1000; i++) {
         try {
-          await throwAttestationError(true);
+          await throwAttestationError();
         } catch (e) {}
       }
-    }
+    },
   });
 
   itBench({
-    id: "Object",
+    id: "Object x1000",
+    runsFactor: 1000,
     fn: async () => {
       for (let i = 0; i < 1000; i++) {
         try {
-          await throwObject(true);
+          await throwObject();
         } catch (e) {}
       }
-    }
+    },
   });
 });
 
-async function throwAttestationError(b: boolean): Promise<void> {
-  if (b) {
-    throw new AttestationError(GossipAction.IGNORE, {
-      code: AttestationErrorCode.UNKNOWN_BEACON_BLOCK_ROOT,
-      root: Buffer.alloc(32),
-    });
-  }
+async function throwAttestationError(): Promise<void> {
+  throw new AttestationError(GossipAction.IGNORE, {
+    code: AttestationErrorCode.UNKNOWN_BEACON_BLOCK_ROOT,
+    root: ZERO_HASH,
+  });
 }
 
-async function throwObject(b: boolean): Promise<void> {
-  if (b) {
-    throw {
-      action: GossipAction.IGNORE,
-      code: AttestationErrorCode.UNKNOWN_BEACON_BLOCK_ROOT,
-      root: Buffer.alloc(32),
-    };
-  }
+async function throwObject(): Promise<void> {
+  throw {
+    action: GossipAction.IGNORE,
+    code: AttestationErrorCode.UNKNOWN_BEACON_BLOCK_ROOT,
+    root: ZERO_HASH,
+  };
 }