**Motivation**

This profile with `--subscribeAllSubnets` flag shows that creating AttestationError takes 4% of the cpu time
<img width="1490" alt="Screen Shot 2021-11-20 at 16 38 10" src="https://user-images.githubusercontent.com/10568965/142721821-ece08a31-5c82-4d2b-9d82-3881b8aa9167.png">

**Description**

Throwing a regular object could improve the speed 500%:
```
test throw AttestationError vs object
    ✓ AttestationError                                                    115.3952 ops/s    8.665875 ms/op        -        197 runs   2.21 s
    ✓ Object                                                              578.6963 ops/s    1.728022 ms/op        -        410 runs   1.21 s

```
+ found from #3416
+ this is another option to improve performance compared to #3402