@dapplion 2021-11-21
@tuyennhv I did some more analysis on this and we should benchmark async operations not async. The difference here is not throw X vs throw Y vs return, but reject X vs reject Y vs resolve.