@tuyennhv 2021-11-24
> more analysis on this and we should benchmark async operations not sync. The difference here is not throw X vs throw Y vs r

@dapplion I just want to show that it takes time to construct LodestarError (4% of the time), if we throw/reject a regular object it's much better