@tuyennhv 2021-12-02
@dapplion the async benchmark test also shows that it's a 5x improvement

```
test throw AttestationError vs object
    ✓ AttestationError                                                    63.24967 ops/s    15.81036 ms/op        -         57 runs   1.41 s
    ✓ Object                                                              315.5363 ops/s    3.169207 ms/op        -       1112 runs   4.04 s
```