@tuyennhv 2021-12-03
@dapplion  latest statistic in my local environment:

```
test throw AttestationError vs object
    ✓ AttestationError x1000                                              88479.32 ops/s    11.30208 us/op        -        123 runs   1.89 s
    ✓ Object x1000                                                        510569.6 ops/s    1.958597 us/op
```