@tuyennhv 2022-04-05
This came at the time when our node had a lot of performance issues, so when we validate gossip messages there are a lot of AttestationErrors. This PR suggested a more performant way to throw errors.

As of now, the performance is so much better so there are so few AttestationErrors, closing for now.