@dapplion 2022-04-05
This PR still shows returning objects vs throwing errors is more efficient. We should eventually adopt that, is there any issue for it?